+++ package.json
@@ -19,7 +19,8 @@
     "react": "^18.2.0",
     "react-datocms": "^3.1.4",
     "react-dom": "^18.2.0",
-    "react-map-gl": "^7.0.19"
+    "react-map-gl": "^7.0.19",
+    "react-multi-carousel": "^2.8.2"
   },
   "devDependencies": {
     "@tailwindcss/forms": "^0.5.3",
